+++ packages/react-components/tests/browser/RichTextInputBoxComponent.spec.tsx
@@ -4,12 +4,13 @@
 import React from 'react';
 import { expect } from '@playwright/experimental-ct-react';
 import { TestRichTextInputBoxComponent } from './TestingComponents/TestRichTextInputBoxComponent';
-import { Locator, Page } from 'playwright-core';
+import { Locator } from 'playwright-core';
 import { test as betaTest } from './FlavoredBaseTest';
 
 const formatButtonId = 'rich-text-input-box-format-button';
 const editorId = 'rooster-rich-text-editor';
 
+// created a separate file for table tests (`RichTextInputBoxComponentTablesTests.spec.tsx`) to speed up the test execution
 betaTest.describe('RichTextInputBoxComponent tests', () => {
   betaTest.skip(({ isBetaBuild }) => !isBetaBuild, 'The tests should be run for beta flavor only');
   betaTest('RichTextInputBoxComponent should be shown correctly', async ({ mount }) => {
@@ -89,93 +90,6 @@ betaTest.describe('RichTextInputBoxComponent tests', () => {
 
     await expect(component).toHaveScreenshot('rich-text-input-box-component-text-formatted-numbered-list.png');
   });
-
-  betaTest('Tables can be added and modified', async ({ mount, page }) => {
-    const component = await mount(
-      <TestRichTextInputBoxComponent disabled={false} minHeight="10rem" maxHeight="10rem" />
-    );
-    await component.evaluate(() => document.fonts.ready);
-
-    await component.getByTestId(formatButtonId).click();
-
-    await component.getByLabel('Insert table').click();
-    await expect(component).toHaveScreenshot('rich-text-input-box-component-insert-table-panel.png');
-    const insertTableOfSelectedSizeButton = page.getByTestId('cell_0_0');
-    await insertTableOfSelectedSizeButton.hover();
-    await expect(component).toHaveScreenshot('rich-text-input-box-component-insert-table-panel-selected-size.png');
-
-    // add table 1*1 to have only one sell for the next step
-    await insertTableOfSelectedSizeButton.click();
-
-    await component.pressSequentially('Test 1');
-
-    await selectTableContextMenu({
-      component,
-      page,
-      menuTitle: 'Insert',
-      subMenuTitle: 'Insert above',
-      text: 'Test 2',
-      cellName: 'Test 1'
-    });
-    await expect(component).toHaveScreenshot('rich-text-input-box-component-table-insert-additional-row-above.png');
-
-    await selectTableContextMenu({
-      component,
-      page,
-      menuTitle: 'Insert',
-      subMenuTitle: 'Insert below',
-      text: 'Test 3',
-      cellName: 'Test 1'
-    });
-    await expect(component).toHaveScreenshot('rich-text-input-box-component-table-insert-additional-row-below.png');
-
-    await selectTableContextMenu({
-      component,
-      page,
-      menuTitle: 'Insert',
-      subMenuTitle: 'Insert right',
-      text: 'Test 4',
-      cellName: 'Test 1'
-    });
-    await expect(component).toHaveScreenshot('rich-text-input-box-component-table-insert-additional-column-right.png');
-
-    await selectTableContextMenu({
-      component,
-      page,
-      menuTitle: 'Insert',
-      subMenuTitle: 'Insert left',
-      text: 'Test 5',
-      cellName: 'Test 1'
-    });
-    await expect(component).toHaveScreenshot('rich-text-input-box-component-table-insert-additional-column-left.png');
-
-    await selectTableContextMenu({
-      component,
-      page,
-      menuTitle: 'Delete',
-      subMenuTitle: 'Delete row',
-      cellName: 'Test 1'
-    });
-    await expect(component).toHaveScreenshot('rich-text-input-box-component-table-delete-row.png');
-
-    await selectTableContextMenu({
-      component,
-      page,
-      menuTitle: 'Delete',
-      subMenuTitle: 'Delete column',
-      cellName: 'Test 2'
-    });
-    await expect(component).toHaveScreenshot('rich-text-input-box-component-table-delete-column.png');
-
-    await selectTableContextMenu({
-      component,
-      page,
-      menuTitle: 'Delete',
-      subMenuTitle: 'Delete table',
-      cellName: 'Test 4'
-    });
-    await expect(component).toHaveScreenshot('rich-text-input-box-component-table-delete-table.png');
-  });
 });
 
 const addList = async (listButtonLabel: string, component: Locator): Promise<void> => {
@@ -193,28 +107,4 @@ const addList = async (listButtonLabel: string, component: Locator): Promise<voi
   await editor.press('Enter');
   await component.getByLabel('Decrease indent').click();
   await editor.pressSequentially('Third line');
-};
-
-const selectTableContextMenu = async ({
-  component,
-  page,
-  menuTitle,
-  subMenuTitle,
-  text,
-  cellName
-}: {
-  menuTitle: string;
-  subMenuTitle: string;
-  text?: string;
-  component: Locator;
-  page: Page;
-  cellName: string;
-}): Promise<void> => {
-  const cell = component.getByRole('cell', { name: cellName });
-  await cell.click({ button: 'right' });
-  await page.getByText(menuTitle, { exact: true }).hover();
-  await page.getByText(subMenuTitle).click();
-  if (text) {
-    await component.pressSequentially(text);
-  }
 };
